improve initial load speed and fix discover feed
there was a bug with the discover feed where the whole feed would re-render when scrolling way down and loading older notes.
