+++ components/DiscoverFeedChips/DiscoverFeedChips.tsx
@@ -1,16 +1,34 @@
 import { Box, Chip } from "@mantine/core";
 import useStyles from "./DiscoverFeedChips.styles";
+import { useEffect, useState } from "react";
+import { NDKEvent } from "@nostr-dev-kit/ndk";
 
 export default function DiscoverFeedChips({
-  labels,
+  events,
   value,
   onChange,
 }: {
-  labels: string[];
+  events: NDKEvent[];
   value: string;
   onChange: (value: string | string[]) => void;
 }) {
   const { classes } = useStyles();
+  const [chipLabels, setChipLabels] = useState<string[]>([]);
+  const chipLabelsHash = chipLabels.join();
+
+  useEffect(() => {
+    const tagNames: string[] = [];
+
+    events.forEach(({ tags }) => {
+      tags.forEach((tag) => {
+        if (tag[0] === "t") {
+          tagNames.push(tag[1]);
+        }
+      });
+    });
+
+    setChipLabels(Array.from(new Set([...chipLabels, ...tagNames])));
+  }, [events, chipLabelsHash]);
 
   return (
     <Box className={classes.box}>
@@ -21,7 +39,7 @@ export default function DiscoverFeedChips({
         position="left"
         className={classes.chipGroup}
       >
-        {["🔥 Latest", ...labels].map((chipName, index) => (
+        {["🔥 Latest", ...chipLabels].map((chipName, index) => (
           <Chip
             key={index}
             value={index === 0 ? "" : chipName}
