+++ components/HomeFeed/HomeFeed.tsx
@@ -10,6 +10,7 @@ export default function HomeFeed() {
   const filter = useMemo<NDKFilter>(
     () => ({
       kinds: [1, 1808 as Kind],
+      limit: 50,
       authors: pubkeys,
     }),
     [pubkeyHash]
